Add validation toast outlet
