+++ src/app/app.component.html
@@ -1,2 +1,3 @@
-﻿<p-toast position="top-right"></p-toast>
+<p-toast position="top-right"></p-toast>
+<p-toast key="validation" position="top-right"></p-toast>
 <router-outlet></router-outlet>
