+++ src/app/feature-module/administration/banks/bank-upsert/bank-upsert.component.html
@@ -62,7 +62,6 @@
             </div>
           </div>
 
-          <app-validation-errors-banner class="mt-3" [errors]="validationErrors()"></app-validation-errors-banner>
 
           <div class="alert alert-danger mt-3" *ngIf="formLevelMessages().length">
             <ul class="mb-0 ps-3">
